Improve notes module.

+++ app/src/main/java/com/elementary/tasks/notes/ImagePreviewActivity.java
@@ -106,7 +106,7 @@ public class ImagePreviewActivity extends ThemedActivity {
 
     @Override
     protected void onDestroy() {
-        if (mItem != null) {
+        if (mItem != null && getIntent().getBooleanExtra(Constants.INTENT_DELETE, true)) {
             RealmDb.getInstance().deleteNote(mItem);
         }
         super.onDestroy();
